Added functional BlendShape importer
There were several issues with the BlendShape importer that I have managed to fix. Primitives were behaving rather oddly and trying to stack BlendShapes leading to errors in the console and failures when loading larger meshes. I believe I have been able to repair it with a rather jank method, but it seems to work on all my current models.